+++ client/components/Checkout.js
@@ -2,7 +2,9 @@ import React from 'react'
 import {CardElement, ElementsConsumer} from '@stripe/react-stripe-js'
 import CardSection from './CardSection'
 import axios from 'axios'
+import OrderSummary from './OrderSummary'
 // import '../../public/common.css'
+//const stripe = require('stripe')('sk_test_4eC39HqLyjWDarjtT1zdp7dc')
 
 class Checkout extends React.Component {
   constructor() {
@@ -25,23 +27,53 @@ class Checkout extends React.Component {
     // We don't want to let default form submission happen here,
     // which would refresh the page.
     event.preventDefault()
+    //const stripe = require('stripe')('sk_test_4eC39HqLyjWDarjtT1zdp7dc')
+
+    /* const paymentIntent = await stripe.paymentIntents.create({
+    amount: 1099,
+    currency: 'usd',
+    // Verify your integration in this guide by including this parameter
+    metadata: {integrationCheck: 'accept_a_payment'},
+  }) */
+    /*  amount: 1099
+canceled_at: null
+cancellation_reason: null
+capture_method: "automatic"
+client_secret: "pi_1GkN7u2eZvKYlo2C7zuu6gdj_secret_3JXbQIMh5QiSbJxGoGluTuB5M"
+confirmation_method: "automatic"
+created: 1589862746
+currency: "usd"
+description: null
+id: "pi_1GkN7u2eZvKYlo2C7zuu6gdj"
+last_payment_error: null
+livemode: false
+next_action: null
+object: "payment_intent"
+payment_method: null
+payment_method_types: ["card"]
+receipt_email: null
+setup_future_usage: null
+shipping: null
+source: null
+status: "requires_payment_method" */
 
     const {stripe, elements} = this.props
+    //const elements = this.props
     console.log('stripe elements', stripe, elements)
     if (!stripe || !elements) {
       // Stripe.js has not yet loaded.
       // Make  sure to disable form submission until Stripe.js has loaded.
       return
     }
-
+    //console.log(paymentIntent)
     const {CLIENT_SECRET} = this.state
+    //const result = await stripe.confirmCardPayment(CLIENT_SECRET,paymentIntent)
     const result = await stripe.confirmCardPayment(CLIENT_SECRET, {
       payment_method: {
         card: elements.getElement(CardElement),
         billing_details: {
-          name: 'Vinayak Khare',
+          name: 'Jenny Rosen',
         },
-        receipt_email: 'khare.vinayak@gmail.com',
       },
     })
     console.log('result', result)
@@ -52,6 +84,7 @@ class Checkout extends React.Component {
     } else {
       // The payment has been processed!
       if (result.paymentIntent.status === 'succeeded') {
+        console.log('The payment has been processed!')
         // Show a success message to your customer
         // There's a risk of the customer closing the window before callback
         // execution. Set up a webhook or plugin to listen for the
